add intro

+++ src/weapon/Arsenal.hpp
@@ -23,7 +23,7 @@ class Arsenal {
 	Weapon& get_weapon_at(int id);
 	Weapon& get_current_weapon();
 	constexpr std::vector<std::unique_ptr<Weapon>>& get_loadout() { return loadout; }
-	[[nodiscard]] auto get_index() const -> size_t { return current_weapon.get(); }
+	[[nodiscard]] auto get_index() const -> size_t { return static_cast<size_t>(current_weapon.get()); }
 	[[nodiscard]] auto size() const -> size_t { return loadout.size(); }
 	[[nodiscard]] auto empty() const -> bool { return loadout.empty(); }
 	bool has(int id);
